// web/src/shared/api/generated/model/mathStepScoreResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface MathStepScoreResult {
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  stepIndex: number | string;
  step: string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  points: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  maxPoints: number | string;
  justification: string;
}